requirements.txt
django>=5.0,<6.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
pydantic>=2.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
markdown>=3.5.0
